refactor CustomizeCssSetting.jsx
